+++ deploy.sh
@@ -1,6 +1,5 @@
 #!/bin/sh
-#lein uberjar
+lein uberjar
 scp target/uberjar/l22.jar app.melt:l22/
 ssh app.melt 'sudo systemctl restart l22'
 ssh app.melt 'systemctl status l22'
-