+++ src/clj/l22/login.clj
@@ -4,8 +4,7 @@
    [l22.db.core :as db]
    [l22.layout :as layout]
    [ring.util.http-response :as response]
-   [struct.core :as st]
-   [taoensso.timbre :as timbre]))
+   [struct.core :as st]))
 
 (def admin-schema
   [[:login
@@ -13,8 +12,8 @@
     st/string
     {:message "Admin only! Your attempt was recorded."
      :validate (fn [login]
-                (let [ret (db/get-user {:login login})]
-                  (and (seq ret) (:is_admin ret))))}]
+                 (let [ret (db/get-user {:login login})]
+                   (and (seq ret) (:is_admin ret))))}]
    [:password
     st/required
     st/string]])
@@ -35,7 +34,7 @@
       (if (and (seq user)
                (:is_admin user)
                (hashers/check (:password params) (:password user)))
-        (-> (response/found "/admin/index")
+        (-> (response/found "/admin/users")
             (assoc-in [:session :identity] (keyword (:login params))))
         (-> (response/found "/login")
             (assoc :flash {:bad "bad password"}))))))