+++ src/clj/l22/password.clj
@@ -1,11 +1,11 @@
 (ns l22.password
  (:require
   [buddy.hashers :as hashers]
+  [clojure.tools.logging :as log]
   [l22.db.core :as db]
   [l22.layout :as layout]
   [ring.util.http-response :as response]
-  [struct.core :as st]
-  [taoensso.timbre :as timbre]))
+  [struct.core :as st]))
 
 (def password-schema
   [[:login
@@ -26,20 +26,19 @@
   (first (st/validate params password-schema)))
 
 (defn password [{:keys [flash] :as request}]
-  (timbre/debug "password flash:" flash)
   (layout/render [request] "password.html"
                  (select-keys flash [:name :message :errors])))
 
 (defn password! [{:keys [params remote-addr]}]
   (if-let [errors (validate-password params)]
     (do
-      (timbre/info "password validation error" (:login params) remote-addr)
+      (log/info "password validation error" (:login params) remote-addr)
       (-> (response/found "/password")
           (assoc :flash (assoc params :errors errors))))
     (let [user (db/get-user {:login (:login params)})]
       (if (hashers/check (:password params) (:password user))
         (do
-          (timbre/info "password changed" (:login params) remote-addr)
+          (log/info "password changed" (:login params) remote-addr)
           (db/update-password!
            (assoc (dissoc params :password :updated_at)
                   :password (hashers/derive (:new-password params))
@@ -47,7 +46,7 @@
           (-> (response/found "/")
               (assoc :flash "password changed")))
         (do
-          (timbre/info "password error" (:login params) remote-addr)
+          (log/info "password error" (:login params) remote-addr)
           #_(layout/render nil "error.html"
                            {:status 404
                             :title "error"