+++ src/clj/l22/routes/home.clj
@@ -9,38 +9,43 @@
    [l22.password :refer [password password!]]
    [ring.util.response]))
 
-(def ^:private version "0.2.14")
-;; only works in development. not in jar.
+(def ^:private version "0.5.2")
+(def ^:private updated_at "2022-07-02 08:29:02")
+;; below only works in development, not in jar.
 ;; (def ^:private version
 ;;   (-> "project.clj" slurp read-string (nth 2)))
 
 ;; cancel by 0.2.14
-#_(defn home-page [{:keys [flash] :as request}]
-    (let [body (-> (client/get "https://w.hkim.jp/loc")
-                   :body
-                   (json/read-str :key-fn keyword))]
-      (layout/render request "home.html"
-                     {:flash flash
-                      :loc (:location body)
-                      :ts (:timestamp body)})))
+;; the reason? cors?
+;; (defn home-page [{:keys [flash] :as request}]
+;;   (let [body (-> (client/get "https://w.hkim.jp/loc")
+;;                  :body
+;;                  (json/read-str :key-fn keyword))]
+;;     (layout/render request "home.html"
+;;                    {:flash flash
+;;                     :loc (:location body)
+;;                     :ts (:timestamp body)})))
 
 (defn home-page [{:keys [flash] :as request}]
   (layout/render request "home.html"
                  {:flash flash}))
 
 (defn about-page [request]
-  (layout/render request "about.html" {:version version}))
+  (layout/render request "about.html" {:version version
+                                       :updated_at updated_at}))
+
+
 
 (defn home-routes []
   [""
    {:middleware [middleware/wrap-csrf
                  middleware/wrap-formats]}
-   ["/" {:get home-page}]
-   ["/about" {:get about-page}]
-   ["/login" {:get  login
-              :post login!}]
-   ["/logout" {:post logout!}]
-   ["/register" {:get  register
+   ["/"         {:get home-page}]
+   ["/about"    {:get about-page}]
+   ["/login"    {:get login
+                 :post login!}]
+   ["/logout"   {:get logout!}]
+   ["/register" {:get register
                  :post register!}]
-   ["/password" {:get  password
+   ["/password" {:get password
                  :post password!}]])